Updated openstack/openstack Project: openstack/horizon  fa2c99d712de6f7003931ed84c4edda5273e6f18
Fix handling of swift object name == pseudo-folder name

Reinstate the proper path for swift pseudo-folders so
that they can be differentiated when adjacent to objects
with the same name.

In adding the test code for this patch the test data
also had to be modified to be more correctly mirroring
the API response data we get.

Change-Id: If0df0be612c0eeb999e2633eb6a00da80b2e0cf8
Closes-Bug: 1596568