Updated openstack/openstack Project: openstack/cinder  b472011f7230925d5dc9d22ac5121d11f4a5cff0
Validate extra_specs type

Currently volume type creation with extra_specs set to null or
a string fails with AttributeError. This is because the current
validation logic assumes that extra_specs is a dictionary. This
patch fixes it by checking whether extra_specs is a dictionary
and raises InvalidInput if it is not.

Change-Id: Icc3580b5b1d784fa9c7fe07ac1c539af1b292075
Closes-bug: #1599106