Updated openstack/openstack Project: openstack/nova  d767164a45051e969d4f1ad7c396611ae6cc00ce
Clean up service version history comments

The service version history comments need to describe what changed
that caused the SERVICE_VERSION to be bumped. This should be an event
like "service does a new thing" or "compute rpc changed". It should not
be a place where we replicate the comment we added to compute/rpcapi.
Several people have just copied the compute-specific comment into this
history, which triggered other people to do the same. It makes it
harder to scan the history and see which version bumps were mechanical,
and which were for more complex events.

This corrects those comments by making them generic, and adds a
bit of help to the block comment above for people that read it.

Change-Id: I4595e72b9de53309a1a6bedc41e7ba31092c580b