Updated openstack/openstack Project: openstack-infra/project-config  18a285882f426e7e7cdfc2413af235c4f573cb23
Add the Tricircle functional test job

The Tricircle provides an OpenStack API gateway and networking automation
to allow multiple OpenStack instances, spanning in one site or multiple
sites or in hybrid cloud, to be managed as a single OpenStack cloud.

As the OpenStack API gateway, the Tricircle should be able to accept
OpenStack API and forward it to regarding bottom OpenStack.

The Tricircle will be installed together with Nova/Cinder/Neutron in
the DevStack, and reuse the tempest use cases to test the Tricircle and
bottom OpenStack which are managed by the Tricircle.

This patch is to add the functional test job in the project-config.
The hook scripts which will be used in the functional test job are
merged in the patch: https://review.openstack.org/#/c/329300/

Change-Id: Ic485432739f9e61242623c0f7b457a80023fe627
Signed-off-by: Chaoyi Huang <joehuang@huawei.com>