Updated openstack/openstack Project: openstack/nova  a92cac899c982eac17d5ef8afbb434487602a37a
Change instance_claim parameter from instance_ref to instance

Now that we're actually passing in an instance object into
instance_claim, let's just change the parameter name to match the
other functions in this file.  Also update the comment to reflect
the fact that it's an object and not a dict.

Change-Id: Ia49c0153d75a434a41e4f865f408330fcd0462e4