Updated openstack/openstack Project: openstack-infra/project-config  a0725ec5f5a4134bb288b7f836ffbe4c5336f766
Add experimental ansible-func tests to openstack-ansible

This patch adds the ansible-role-jobs job template to the
openstack/openstack-ansible repository in order to enable
tox-based functional tests on multiple platforms.

As the ansible-role-jobs job template also contains the
docs/linters/announce-release jobs, so they are removed
from the jenkins jobs list for the repo.

Change-Id: Ibf40baabf489a11f21f1dcc5142a1452f68ccdba