Updated openstack/openstack Project: openstack/horizon  0d5d14763bd9692be2a2ce7f985176776abded6e
Modified modal error message to display properly when present

To recreate the issue, using the Angular Create Image, create an image
with Name 'asdf', Location 'asdf', and Format 'raw' (or similar).  Before
this patch, you see an odd '[Object] [Object]' notice; after this patch
it displays some actual information.

Honestly, I'm not sure we want this display as I'm not sure it's good UX.
But it is a feature we have right now.

Change-Id: I9a214d12853e2884c1046308f253e0337f5c9032
Closes-Bug: 1561479