Updated openstack/openstack Project: openstack/keystone  b5b7c692b9b04ace7683df9da3d4f1f8801c01c1
Pass request to build_driver_hints

Pass the request object through to the build_driver_hints function so
that it can use request.params instead of context.query_string.

This shows the problem with the domain_id filter in list roles. For
whatever reason the driver doesn't filter on domain_id=None by default
so we need to add this filter manually. Because we can no longer
influence the query string we add it to the hints object directly if the
param is not set.

Change-Id: I732c603a24f6b884820ee6837d4c0f752d77987d