Updated openstack/openstack Project: openstack/python-cinderclient  dd19ccd86d9a09329ae8270f7c6ff5efc108b8aa
Fix batch deleting issue in volume_type.unset_keys()

cinderclient volume_type.unset_keys() only delete the first key even if
multiple keys are applied. The response of manager._delete() is tuple object,
it is not "None" if the deleting success, so the batch operation would
be broken when the first key is deleted successfully. The patch fix this
issue.

Change-Id: I60378a32cdc52aacdf869d69b246dec7eb6cdb77
Closes-Bug: #1596511