Updated openstack/openstack Project: openstack/oslo.db  58d60ca5b2465c96adec36c286328b1b8b8b512f
exception: make message mandatory in DbMigrationError and deprecates it

An exception message with None as an error message is not really useful. Since
DbMigrationError is also wrongly written with a lower case b whereas everything
else has a major B, let's create a new correct class and use it.

debtcollector is used to deprecate the old one, and inheritance makes sure the
current code works.

Change-Id: Id9a477aff38afd5b664b75a549ad561f4c24b6f0