Updated openstack/openstack Project: openstack/nova  bcabd1a5b1d9ba6a40ed16673e124ab6436c1f0e
Prevent instance disk overcommit against itself

The CPU/RAM filters currently prevent scheduling an
instance that is larger than the hypervisor total resource size.
This does the same for the disk filter.

Previously it was possible to schedule a 10Gb instance
on an 8Gb host if the overcommit ratio permitted.

Change-Id: Ia9963434bfb8845903b4e138545660a3827418e9