Updated openstack/openstack Project: openstack/tripleo-heat-templates  f91f08a5288820a77f45099092914a65a038caa5
Parametrize and increase default of swift-proxy node_timeout

We've had problems uploading images on slow virtualized environments,
where the requests to swift would time out. The issue has been seen both
on undercloud and overcloud. The default timeout is 10 seconds, the
issue has been reportedly reproduced on undercloud with node_timeout as
high as 30 seconds, but not with 60 seconds yet. Set the default timeout
to 60 seconds on overcloud too.

Change-Id: I7d486cf4dc9768ddbf71ab71e92db8d2ef29978e
Closes-Bug: #1594725