Updated openstack/openstack Project: openstack-infra/storyboard  04f5c7c56f522fc20eab69a038aa2973e1e278cc
Add a paragraph instructing api users how to get a token

First time api users do not know how or where to obtain a user
token. A user has to use the gui to get an initial token to use
the api, including the user token api, which will allow them to
get subsequent tokens.

Change-Id: Ib1696e60be02c17aa1a4020294fc27eb1a57b977