Updated openstack/openstack Project: openstack/tempest  12f8d4a190ec53754eef19ec7434b8fbd52aedaa
Add inherited role assignments tests

Inherited role assignments are really useful when dealing with
hierarchies - we must test that granting an inherited role in a higher
level of the hierarchy has effect in the project's tree.

The verification is done via the "role_assignments" API and its query
parameters, such as "effective" (which lists the role assignments that
take effect in each project layer).

Change-Id: I0f5a54d85a28c8c011fd3489cf55629b9caa1dcf