Updated openstack/openstack Project: openstack/keystone  d03a84ea8b59a35fc8aacf497845fdb2d89e7093
Refactor: [ldap] suffix should not be an instance attribute

In investigating the usage of the [ldap] suffix option, I noticed that
it's only used once, and there was no reason for it to become an
instance attribute. This tiny refactor clarifies exactly where and how
it's used.

I'd further recommend that this entire configuration option be
deprecated in favor of the [ldap] user_tree_dn and [ldap] group_tree_dn
options.

Change-Id: I7c7dacbc65aa876d85af292e916125e7fc6964b9