Updated openstack/openstack Project: openstack/cinder  61602912f49afc8c68cbd15859339c8692f4f098
Moving the OPTIONAL_FIELDS inside the OVO  object

In working patch 251989, Michal Dulko points out that
all the OVOs should have OPTIONAL_FIELDS inside object.

This patch will fix the issue in OVO, CGSnapshot.

Now that the OPTIONAL_FIELDS is part of the class definition,
_from_db_object should no longer be a staticmethod,
but should be a classmethod.

Change-Id: I9603cc3839dfd6e1cf102a77aeed3e9edc661bb4