Updated openstack/openstack Project: openstack/ironic-python-agent  185780822f38603aec93653a75f7161bfc8b2931
Bump CoreOS to 1010.6.0 (last current stable)

This approach was used instead of fetching the latest version number at
build time because that meant a CoreOS update could break our gates at
any given time.

A new CoreOS make target, 'pin_latest', was made available. This target
bumps the "pinned" version to the latest. An eventual proposal bot will
call this target to ensure we stay up to date. This also offers the
possibility for downstream operators that want to build the latest
version to call this target before building. This leads to the same
result as fetching directly in the python build script, but with more
flexibility.

This is a second attempt to merge this change, as before it caused
frequent CI failures/timeouts. The expectation is that with CoreOS only
gating on IPA, this would be lower impact and easier to mitigate.

Change-Id: I78cb2efff2293b1591b30b7e39df9fc2af98d0b2
Closes-bug: 1581673
Co-Authored-By: Arun S A G <sagarun@gmail.com>
Co-Authored-By: Jim Rollenhagen <jim@jimrollenhagen.com>