Updated openstack/openstack Project: openstack/swift  fa1f67b116e8712e2235c10c9166834d1151d0fa
Add ringbuilder tests for --yes option

Also added a Timeout class to test.unit to wrap possible long-running
functions. For example, if there is some regression and the "--yes"
argument is no longer evaluated correctly and the test excepts some
keyboard input, it will be terminated after a few seconds to ensure
there is no long-running blocker on the gate.

Change-Id: I07b17d21d5af7fcc594ce5319ae2b6f7f58df4bb