Updated openstack/openstack Project: openstack/swift  1ac9ceb548948c96c16108c17ca7c88b32f703fa
Add description of server options (all, main, rest) for swift-init help

Running swift-init with -h, --help, or no arguments
displays help for the command. The help does not
document the 'main', 'all', and 'rest' options.
These are documented in the man page.

This patch adds all these server options in the
help of swift-init.

Change-Id: I8e27589912ae72ace14c955e66b86942bc23d9f7
Closes-Bug: #1580722