Updated openstack/openstack Project: openstack/keystone  99344c4644cdf90d004ea1f02975ec239a1a56f1
Ensure status code is always passed as int

There is some inconsistency in calls to the wsgi render_response
function. Sometimes the status tuple uses an int for the http
status code and other times a string. The Python 3.5 gate job
discovered the problem.

This patch normalizes all calls to use an int.

Change-Id: I136b01f755ff99dfba244e79068fdaae614b2091
Closes-Bug: #1599983