Updated openstack/openstack Project: openstack/keystone  3e07bf6c57f759271ff5aab78e540b4d90c4045f
Move the auth plugins abstract base class out of core

This patch moves the auth plugins abstract base class out of core and
into plugins/base.py

This removes dependencies where backend code references code in the
core. The reasoning being that the core should know about the backend
interface, but the backends should not know anything about the core
(separation of concerns). And part of the risk here is a potential for
circular dependencies.

Partial-Bug: #1563101

Change-Id: I4413ef01523d02c30af97e306069229252cb4971