Updated openstack/openstack Project: openstack/keystone  1cae3c42ee57b3eac2d006a2af84978d152902c1
Expose bug with Fernet tokens and trusts

When a trustee has a trust established with a project in a different domain, it
is possible for trust-scoped tokens to still be valid even though the trustee's
domain has been disabled. This commit adds a test to expose the behavior when
using the Fernet token provider.

Change-Id: I6e287735168ecdd52c28b7e5acf010b7a82e33c7
Partial-Bug: 1532280