Updated openstack/openstack Project: openstack/nova  976b85faf7b673e4964c7f47adde37f8e0668e56
Add Allocation and AllocationList objects

An Allocation represents a use of some class of resource from a
particular resource provider, by a consumer. They can be created
and destroyed.

It is also possible get a list of all allocations associated with a
particular resource provider.

This also includes: a small correction to functional tests of Inventory
so that its structure (where resource_class is defined) is aligned
with the Allocation tests.

Co-Authored-By: Dan Smith <dansmith@redhat.com>
Partially-Implements: blueprint resource-providers-allocations
Change-Id: Iaec8a8e318adfd8feae6496ad962bc02d13adebd