Updated openstack/openstack Project: openstack/keystone  28d86afb88da2d1efb8ffc63260a80ed9e12cd84
Move the trust abstract base class out of core

This patch moves the trust abstract base class out of core and
into backends/base.py

This removes dependencies where backend code references code in the
core. The reasoning being that the core should know about the backend
interface, but the backends should not know anything about the core
(separation of concerns). And part of the risk here is a potential for
circular dependencies.

Partial-Bug: #1563101

Change-Id: Ibe9203635ee4ad5d1afee3ad751f04de6ef66f94