Updated openstack/openstack Project: openstack/keystone  7faa001f10b52e8363494eaeda88cadf1cb91093
Move the credential abstract base class out of core

This patch moves the credential abstract base class out of core and
into backends/base.py

This removes dependencies where backend code references code in the
core. The reasoning being that the core should know about the backend
interface, but the backends should not know anything about the core
(separation of concerns). And part of the risk here is a potential for
circular dependencies.

Partial-Bug: #1563101

Change-Id: I4007a5fe05e61f08999a4b3f6801727c08aaaa14