Updated openstack/openstack Project: openstack/sahara  82e87ab0e6ebbd5af170b36dd45b838c53757eb9
allow to specify notifications transport url

new feature was introduced in oslo.messaging (in Mitaka)
which allows to specify transport url for notifications
purpose. so, now setup of transports will be following:
we will try to setup transport for notifications
using this new feature, and if such setup is failed,
we will fallback to old scheme with re-using
service transport for notifications (for
backward compatibility).

Change-Id: I251d465863c33f60bb3b3850e987238bd326c461
Closes-bug: 1579945