Updated openstack/openstack Project: openstack-infra/project-config  22a8e0c41f14e274b1dc2c4fc62742e871489c8f
puppet-ceph: move puppet4 jobs into p-o-i

Like other Puppet OpenStack modules, run puppet4 jobs as -nv in
puppet-openstack-integration so we don't run it in puppet-ceph check
pipeline. When we'll decide that Puppet4 is stable for us, we'll switch
every Puppet4 job to the gate but it's not the case yet.

Change-Id: I121d4b72d8e64ee136ecd5e095baa7e45c8f9eaa