Updated openstack/openstack Project: openstack/heat  0bdca4c3c9d3f4b4122265edef0483724ab401f0
Improve validation for some functions

Move validation code to __init__ method for these list of
functions: Removed, Repeat, Yaql, so now validation error will
be found during template.parse() and returned with a path
to the function in template. Don't change validation for
GetAttr, because it resolves some arguments and we need to parse
template before we can use resolve().
Also, remove check for Removed function in translation, because
now we can't have Removed in properties.

Change-Id: I4f2e1ebee37e82badcdca9110cadc649aa01be8c