Updated openstack/openstack Project: openstack-infra/storyboard  4826d1905dea9a31ffe42f5278f325500c59ac42
Add example commands for the Project api

Currently the api documentation does not include example commands.
It would be very friendly for our users to have some example commands
to follow and use the api.

This patch adds examples to the Project section and the Project
Groups section of the api documentation.

Co-Authored-By: Adam Coldrick <adam.coldrick@codethink.co.uk>
Change-Id: I24fbf6f0fb962698af1ba8415504dc4c2b903f7f