Updated openstack/openstack Project: openstack/tripleo-heat-templates  986c983abc6751243183bca252149f8085a4ca84
Pass MysqlVirtualIP without brackets to EndpointMap

While the endpoints do need brackets around IPv6 addresses these
are not wanted by some of the puppet classes so we must pass the
non bracketed version as well.

It will allow us to remove the need for the MysqlVirtualIP param
passed to the controller role thouh when hieradata/database.yaml
is emptied.

Change-Id: If264b02a134b96368035f032e05d02e84f6499ed