Updated openstack/openstack Project: openstack/tripleo-puppet-elements  ea61150ddeed4578c28b497ea057dcdc53f90662
Install crudini in overcloud again

We had crudini installed before due to some dependency probably, but
either we're no longer installing the package that depended on crudini,
or that dependency was removed, so we need to pull in crudini explicitly
to make sure we can still use it.

Our overcloud upgrade scripts rely on the crudini tool to perform
one-off config tweaks during major upgrades.

Change-Id: I8244ed3199a880b8c26a3e43c80728e650c3ad20