Updated openstack/openstack Project: openstack/nova  92134b4c1806f212a592a2aec69819e075fb45a2
Added support for new block device format in vmops

Modified vmops methods to make use of the new block device format.

This new format allows for more flexibility when using block device,
such as the ability to select the bus for the attachment and the boot
index.

Co-Authored-By: Claudiu Belu <cbelu@cloudbasesolutions.com>

Partially implements: blueprint hyper-v-block-device-mapping-support

Change-Id: Ib6ff12b2710143798c1c29376edd9c1364d451b9