Updated openstack/openstack Project: openstack-infra/storyboard  974bd9522d49dc731f91a3ad81dfeeca85073ce2
Use a common requirements file

We can use newer setuptools and pbr to use environment markers in our
requirements files which allows us to select deps if the environment
matches. With this we restrict the email dep to python < 3.0 which is
the only dep we have that doesn't work across python 2 and 3.

Note that this also updates the pbr machinery to pull in this new
feature support. Part of this update requires updating hacking to avoid
conflicting pbr requirements. This in turn has created some churn in the
format of the code but should largely be a noop.

Change-Id: I0be5dd8a6b33a51329077b5a5f4c7f5576829956