Updated openstack/openstack Project: openstack/neutron  6087f1d12179f7f9bf1969fe94f96d638dfb99ce
Extension to tell when deferred binding is in effect

This extension exists mainly to let Nova know when deferred port
binding is in effect and its okay for the port to not have any IP
addresses.

The extension is implemented by the segments service plugin with a
very simple implementation to begin with.  This implementation is
likely to evolve.

Change-Id: I591d32df512712e4de36fe20e389b3a14d58157f
Partially-Implements: blueprint routed-networks