Updated openstack/openstack Project: openstack/nova  232ffaea589cb88fcd90fd90e36056da1826ee3b
Revert "Detach volume after deleting instance with no host"

This reverts commit 5ce74fa06c0e7a70fdc927b2c1f364af83f7de1d.

We think this is causing a race and the postgres job to fail
since it's erroneously doing local deletes and not cleaning
up from the computes. We're not totally sure why it would
only impact the postgres job though, but the original change
failed the job and was rechecked, and the time it was merged
coincides with when the postgres job started spiking with
failures.

Related-Bug: #1600005
Closes-Bug: #1600031

Change-Id: I0ed184b579b8a8d861e4d2a7c317bf0bc0623d50