Updated openstack/openstack Project: openstack/keystone  005fd8a78cc6a5a37dbcfa24bdab2d5af80abc2b
Add the oslo_context to the environment and request

So in the effort to standardize our usage of components add the
oslo_request object to the request so that we can actually use it.

We subclass request as there will be information there that we will want
to keep in addition to the standard auth params.

Whilst there seems to be a lot of contexts floating around this is
transitional. The split will end up very similar to today except instead
of an unstructured context_dict we will have a webob.Request for request
related params, and instead of an unstructured auth_context dict we will
have an oslo_context.RequestContext.

Change-Id: I1a96233ee3deebdb143645b4be3a7b4068b2b165