Updated openstack/openstack Project: openstack/cinder-specs  bf4ec715e8c39829bf04d46d7201cb58635e2f36
Generic Volume Group

This spec designs a generic volume group for various
group operations. A group type will be introduced to
describe a group, similar to a volume type to a volume.
Basic APIs including create, delete, update, list,
and show will be introduced for the generic volume group
in this spec.

Other features can be added to the group dependent on
the basic group functionalities in this spec.

Change-Id: I89eb60f80e846a30cfcde910a775ed42288bfccd
blueprint: generic-volume-group