Updated openstack/openstack Project: openstack/cinder  7e6f4f2ed0623353a88f67a0879f6519bb45926c
Fix failure with "None" volume type in Pure drivers

We were not checking if the type id was None before calling
the util method to get the volume type. Turns out the helper
will raise an exception if the type is None (intentionally).

This fixes the issue by checking first in the driver before
using the util method. Added unit test coverage for this
scenario as well.

Change-Id: Ie3c7b56fafb4a8c07853a9c1cd11e7e26444bdc1
Closes-Bug: #1600362