Updated openstack/openstack Project: openstack/cinder  a76f4a1b942df0f37d0cf8d126faa27f77722bb0
Add strict Boolean checking for volume manage

There is no strict boolean checking for the parameter
"bootable" of API /os-volume-manage, so that any invalid
boolean value can be specified.
This patch adds a strict checking for it to prevent
invalid value, and adds a test for this change as well.

Change-Id: I0d79a0bb173aaeeea0fe6d735213c70c109ccd69
Partial-Bug: #1594261