Updated openstack/openstack Project: openstack/horizon  37df1311266482dba8369a688c3b396d459fb93b
Magic-Search isn't rendered unless config is ready

This patch alleviates a problem where sometimes the magic-search element
needs to be rendered, but the configuration isn't ready yet.  This simply
skips the rendering unless the primary configuration parameters are in place.

This is evident when Launch Instance is launched via an item action, such as
within the Image items.

Change-Id: I10452428e2d17cd17c98c65cbf8147dc7ab14621
Closes-Bug: 1600352