Updated openstack/openstack Project: openstack-infra/gear  defb9722b8188981f15555ba04d5598da722e9c7
Support TCP keepalives in geard

Support configuration of TCP keepalives in gear server.
When connections are closed not gracefully they stays opened
for a long time, which leads to file descriptors leak.
Default values for interval, probe and count are such as in
most Linux distributions.
By default is disabled to keep compatibility with older versions.

Change-Id: I335f68a24dda409b1a48c6e4e520ec08dfd38078