Updated openstack/openstack Project: openstack/cinder-specs  bb235f2104686f71fc94ecb32711aff5f3355f6a
Group Snapshots

This spec is dependent on the generic volume group spec.
The purpose of proposing this spec is to migrate the
existing consistency groups support to use the generic
volume group constructs and also provide a common API
for group snapshots that works for all storage backends.

The generic volume group spec is here:
https://review.openstack.org/#/c/303893/

Partial-Implements blueprint: generic-volume-group

Change-Id: I73be6a4e81358c68a14c2d122396e44836bac756