Updated openstack/openstack Project: openstack/cinder  c2f7e72e86d8792463402d1c5f7e22a61cf4b71a
Send scaleio id to os-brick

Send the volume's provider_id attribute to os-brick for
attaching volume. Using the provider_id is safer than using the
encoded name from the volume id.

I will submit a patch in the connector in os-brick to use the
provider_id once this patch is merged.

Change-Id: I1ccc94b2bc042e65a7aec315e322cc65bee52c54